% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{assignMetaData}
\alias{assignMetaData}
\title{assignMetaData}
\usage{
assignMetaData(
  seurat_obj,
  samplesheet,
  multiplex = c("none", "hash", "demuxlet"),
  barcode_df = NULL
)
}
\arguments{
\item{seurat_obj}{Seurat object for a sample to be annotated with metadata.}

\item{samplesheet}{Samplesheet containing per sample data to be added.}

\item{multiplex}{Character vector indicating what kind of multiplexing, if any, should be used. "none" assumes no multiplexing is occuring. "hash" looks for hashtags in the samplesheet and adds them according to exisiting metadata in the object for cell hashing. "demuxlet" indicates that samples have been demuxed in some other fashion (e.g. by genetics) and you are providing a dataframe indicating which cells are assigned to which identity.}

\item{barcode_df}{A dataframe containing barcode-sample assignments, such as the \code{[prefix].best} file generated by demuxlet.}
}
\value{

}
\description{
Adds metadata to a seurat object from a samplesheet. Does this if the samples are singular or contain multiplex data. Returns a seurat object with metadata.
}
