% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calcGEMClassification}
\alias{calcGEMClassification}
\title{calcGEMClassification}
\usage{
calcGEMClassification(list, primary_species_prefix, secondary_species_prefix)
}
\arguments{
\item{list}{A list of gene expression matrices to feed the function}

\item{secondary_species_prefix}{}
}
\value{

}
\description{
calcGEMClassification
}
