% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{constructGEMs}
\alias{constructGEMs}
\title{constructGEMs}
\usage{
constructGEMs(
  sample = NULL,
  gem_matrix,
  adt_matrix = NULL,
  norm_adt_matrix = NULL,
  hashtags = FALSE
)
}
\arguments{
\item{sample}{Character string to append onto each cell barcode.}

\item{gem_matrix}{Matrix containing gene expression data only. Assumes that
rows are genes and that columns are cells.}

\item{adt_matrix}{Matrix containing antibody capture data. If NULL, assumes
that no antibody capture data exists.}

\item{hashtags}{Character vector containing cell hashing tags (e.g. "HTO1")
found in the antibody capture data. If FALSE, assumes that no cell hashing
has taken place.}
}
\value{

}
\description{
Constructs the Gene Expression Matrices (GEMs) with the given
parameters for multimodal data. Returns a final seurat object.
}
