% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{FindPCAElbow}
\alias{FindPCAElbow}
\title{FindPCAElbow}
\usage{
FindPCAElbow(
  object,
  assay = "RNA",
  slot = "data",
  reduction = "pca",
  name = paste0("max_", reduction, "_dim"),
  n_pcs = 100,
  elbow_th = 0.025,
  perform_new_pca = FALSE
)
}
\arguments{
\item{object}{Seurat object}

\item{assay}{Assay to examine}

\item{slot}{Slot to use}

\item{reduction}{Dimensional reduction to look for (e.g. "pca" or "ica")}
}
\value{
A copy of the object with the
largest significant principal component stored
as \code{max_pca_dim} in the object@misc slot
}
\description{
Identify the upper bounds of
significant principal components
}
