% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getClosestFactors}
\alias{getClosestFactors}
\title{getClosestFactor}
\usage{
getClosestFactors(int)
}
\arguments{
\item{int}{A positive integer as a numeric value.}
}
\value{

}
\description{
Gets the closest factor pair for an integer
}
