% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{GroupSmallClusters}
\alias{GroupSmallClusters}
\title{Title}
\usage{
GroupSmallClusters(
  ids,
  SNN,
  threshold = 20,
  group.singletons = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
Title
}
