% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{HTODemux2}
\alias{HTODemux2}
\title{HTODemux2}
\usage{
HTODemux2(
  object,
  assay = "HTO",
  positive.quantile.range = seq(0.8, 0.95, 0.05),
  init = NULL,
  nstarts = 100,
  kfunc = "clara",
  nsamples = 100,
  seed = 42,
  verbose = TRUE
)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
HTODemux2
}
