% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{PurgeXenoGenes}
\alias{PurgeXenoGenes}
\title{PurgeXenoGenes}
\usage{
PurgeXenoGenes(object, primary_species_prefix, secondary_species_prefix)
}
\arguments{
\item{object}{Formerly mixed species Seurat object}

\item{primary_species_prefix}{Literal string or regex prefix denoting
genes from the species of interest}

\item{secondary_species_prefix}{Literal string or regex prefix
denoting genes from the species used as control}
}
\value{
Seurat object
}
\description{
Removes counts and gene loadings for a secondary species and strips the primary species
prefix from the remaining genes.  Useful when cleaning an object that had controls cells added for
multi-modal assays such as CITE-seq or REAP-seq.
}
\examples{
pbmc <- PurgeXenoGenes(object = pbmc, primary_species_prefix = "hg38-", secondary_species_prefix = "mm10-")
}
