% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ScoreIgDiffs}
\alias{ScoreIgDiffs}
\title{Title}
\usage{
ScoreIgDiffs(object, assay = "RNA", ig.hc, ig.lc, conserve.memory = FALSE)
}
\arguments{
\item{conserve.memory}{}
}
\value{

}
\description{
Title
}
